# contracts/console-api.openapi.yaml
openapi: 3.0.3
info:
  title: Keel 控制台 API
  version: 1.0.0-draft
  description: |
    keel-console 九个页面消费的全部接口，由 keel-server 实现。

    这份文件是前端、mock 数据、后端实现三方的唯一事实来源：
    - 前端只调这里定义的接口，不直接调 Langfuse / 薄网关 / Prometheus
    - mock 数据由这份文件生成，接口切真时前端只改 baseURL
    - 后端 P1 按这份文件实现，字段不一致以这里为准

    和 keel/v1 五份契约的关系：那五份是智能体侧协议（/v1/invoke、审计事件、
    追踪属性、manifest、评测集），这份是控制面 REST API，两者独立演进。
    引用到的枚举值必须和 contracts/ 下对应 schema 保持一致。
servers:
  - url: https://keel.example.com/api/v1
    description: prod
  - url: http://localhost:5173/api/v1
    description: 本地 mock（MSW 拦截）
security:
  - bearerAuth: []
...
paths:
...
  /approvals:
    get:
      tags: [approvals]
      operationId: listApprovals
      summary: 审批单列表
      description: 审批人只看得到分给自己的单子，由后端按当前用户过滤。
      parameters:
        - $ref: "#/components/parameters/Env"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ApprovalStatus" }
        - name: agent
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PageMeta"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/Approval" }
    post:
      tags: [approvals]
      operationId: openApproval
      summary: 发起审批单
      description: |
        导出、工具变更、智能体配置与下线都从这里建单。runId 非空时必须已经有对应的 agent_run。
        命中策略冷却期时不新建单，直接返回上一张 APPROVED。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subjectType, subjectRef, summary, actorUser]
              properties:
                subjectType: { type: string, enum: [tool.call, tool.config, agent.config, agent.retire, data.export] }
                subjectRef: { type: string }
                summary: { type: string }
                actorUser: { type: string }
                agent: { type: string }
                risk: { $ref: "#/components/schemas/Risk" }
                runId: { type: string, nullable: true }
                traceId: { type: string }
                payloadDigest: { type: string }
                policyName: { type: string, nullable: true }
                env: { $ref: "#/components/schemas/EnvName" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Approval" }
        "400":
          description: 参数不合法
          $ref: "#/components/responses/Error"
        "404":
          description: 策略、智能体或 run 不存在
          $ref: "#/components/responses/Error"
  /approvals/{id}/decision:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [approvals]
      operationId: decideApproval
      summary: 批准 / 驳回
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [APPROVE, REJECT] }
                comment: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Approval" }
        "409":
          description: 单子已过期或已被处理
          $ref: "#/components/responses/Error"
  /runs:
    get:
      tags: [approvals]
      operationId: listSuspendedRuns
      summary: 挂起等人回话的执行
      description: 只返回 status=SUSPENDED 且 reason 为 input_required / handoff 的 run；reason=approval 的在 /approvals 里。
      parameters:
        - $ref: "#/components/parameters/Env"
        - name: agent
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PageMeta"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/SuspendedRun" }
    post:
      tags: [approvals]
      operationId: openSuspendedRun
      summary: 登记一次等人回话的挂起
      description: reason 只接受 input_required 或 handoff。同一个 runId 再次登记且仍是同一挂起原因时返回原记录。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [runId, agent, reason, prompt, actorUser]
              properties:
                runId: { type: string }
                agent: { type: string }
                env: { $ref: "#/components/schemas/EnvName" }
                traceId: { type: string }
                reason: { type: string, enum: [input_required, handoff] }
                prompt: { type: string }
                actorUser: { type: string }
                deadline: { type: string, format: date-time, nullable: true }
                checkpointRef: { type: string, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/SuspendedRun" }
        "400":
          description: 参数不合法
          $ref: "#/components/responses/Error"
        "404":
          description: 智能体不存在
          $ref: "#/components/responses/Error"
        "409":
          description: runId 已存在且不能再次挂起
          $ref: "#/components/responses/Error"
  /runs/{runId}/input:
    parameters:
      - name: runId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [approvals]
      operationId: answerSuspendedRun
      summary: 回复挂起的执行。keel-server 据此调智能体的 /v1/runs/{id}/resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "409":
          description: 执行已不处于可恢复状态（RUN_NOT_RESUMABLE / RUN_EXPIRED）
          $ref: "#/components/responses/Error"

  # ---------------------------------------------------------------- 审计
  /audit/events:
    get:
      tags: [audit]
      operationId: listAuditEvents
      summary: 审计事件查询
      parameters:
        - $ref: "#/components/parameters/Env"
        - name: agent
          in: query
          schema: { type: string }
        - name: risk
          in: query
          schema: { $ref: "#/components/schemas/Risk" }
        - name: action
          in: query
          schema: { $ref: "#/components/schemas/AuditAction" }
        - name: decision
          in: query
          schema: { $ref: "#/components/schemas/AuditDecision" }
        - name: traceId
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PageMeta"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/AuditEvent" }
  /audit/verify:
    post:
      tags: [audit]
      operationId: verifyAuditChain
      summary: 手动触发哈希链校验
      parameters:
        - name: agent
          in: query
          schema: { type: string, description: 缺省校验全部智能体 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          checked: { type: integer, description: 校验条数 }
                          intact: { type: boolean }
                          brokenAt:
                            type: string
                            nullable: true
                            description: 断链处的 eventId
                          elapsedMs: { type: integer }
  /audit/exports:
    post:
      tags: [audit]
      operationId: requestAuditExport
      summary: 申请导出（自动建审批单，批准后才生成脱敏文件）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filter]
              properties:
                filter:
                  type: object
                  additionalProperties: true
                  description: 与 /audit/events 相同的筛选条件
                reason: { type: string }
      responses:
        "202":
          description: 已建审批单
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          exportId: { type: string }
                          approvalId: { type: string }

  # ---------------------------------------------------------------- 模型网关
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: auth-gateway 签发，aud=keel-api
  parameters:
    Env:
      name: env
      in: query
      schema: { type: string, enum: [all, dev, test, staging, prod], default: all }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, enum: [10, 20, 50, 100], default: 10 }
...
  responses:
    Error:
      description: 统一错误体
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    Envelope:
      type: object
      required: [code, traceId]
      properties:
        code:
          type: string
          description: 成功固定为 OK，失败为 contracts/error-codes.yaml 里的错误码
          example: OK
        message: { type: string, nullable: true }
        traceId: { type: string }
        data: {}
    ApiError:
      type: object
      required: [code, message, traceId, retryable]
      properties:
        code: { type: string, example: TOOL_NOT_GRANTED }
        message: { type: string }
        traceId: { type: string }
        retryable: { type: boolean }
    PageMeta:
      type: object
      required: [page, size, total]
      properties:
        page: { type: integer }
        size: { type: integer }
        total: { type: integer }
    EnvName:
      type: string
      enum: [dev, test, staging, prod]
    Risk:
      type: string
      enum: [LOW, MID, HIGH]
...
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, EXPIRED]
    AuditAction:
      type: string
      description: 与 contracts/audit-event.schema.json 的 action 枚举完全一致，两边必须同时改。
      enum:
        [
          invoke,
          tool.call,
          sql.execute,
          approval,
          config.change,
          data.export,
          run.suspend,
          run.resume,
          agent.register,
          release.gate,
        ]
    AuditDecision:
      type: string
      enum: [allowed, denied, pending, approved, rejected]
...
    Approval:
      type: object
      properties:
        id: { type: string, example: ap_0912 }
        subjectType:
          type: string
          description: 草案（P3-1 定稿）。与 approval_request.subject_type 一致，见 docs/specs/p0/P0-1a-run-lifecycle.md
          enum: [tool.call, tool.config, agent.config, agent.retire, data.export]
        subjectRef: { type: string, description: 工具名 / 智能体名 / 导出申请 id }
        runId: { type: string, nullable: true, description: 非空表示有一次执行正挂起等这张单 }
        policyName: { type: string, nullable: true, example: 班组长审批 }
        agent: { type: string }
        tool: { type: string, description: 仅 subjectType=tool.* 时有值；新代码用 subjectRef }
        risk: { $ref: "#/components/schemas/Risk" }
        traceId: { type: string }
        actorUser: { type: string }
        summary:
          type: string
          description: 人可读的动作摘要；不含完整 payload，原文需另走审批接口
        payloadDigest: { type: string }
        status: { $ref: "#/components/schemas/ApprovalStatus" }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
        decidedBy: { type: string, nullable: true }
        decidedAt: { type: string, format: date-time, nullable: true }
    SuspendedRun:
      type: object
      description: 草案（P3-1 定稿）。挂起等人回话的执行，没有审批单；审批中心「人工介入」一组的数据来源
      properties:
        runId: { type: string, example: r_7b4a }
        agent: { type: string }
        env: { $ref: "#/components/schemas/EnvName" }
        traceId: { type: string }
        reason: { type: string, enum: [input_required, handoff] }
        prompt: { type: string, description: 要问用户的话 / 交接说明 }
        actorUser: { type: string }
        createdAt: { type: string, format: date-time }
        deadline: { type: string, format: date-time, nullable: true }

    # ------------------------------------------------------------ 审计
    AuditEvent:
      type: object
      description: 与 contracts/audit-event.schema.json 同构，额外带 hash、prevHash、hashVerified 供详情展示哈希链
      properties:
        eventId: { type: string }
        ts: { type: string, format: date-time }
        agent: { type: string }
        env: { $ref: "#/components/schemas/EnvName" }
        traceId: { type: string, nullable: true }
        actor:
          type: object
          properties:
            userId: { type: string }
            orgId: { type: string, nullable: true }
            role: { type: string, nullable: true }
        action: { $ref: "#/components/schemas/AuditAction" }
        resource: { type: string }
        risk: { $ref: "#/components/schemas/Risk" }
        decision: { $ref: "#/components/schemas/AuditDecision" }
        approver: { type: string, nullable: true }
        payload:
          type: object
          additionalProperties: true
          description: 只含 manifest captureFields 白名单字段，已脱敏
        inputDigest: { type: string }
        hashVerified: { type: boolean }
        hash: { type: string, description: 本事件 sha256 hex }
        prevHash: { type: string, nullable: true, description: 同智能体链上一条的 hash，链头为空 }

    # ------------------------------------------------------------ 模型网关
...